add errorMessage component
