+++ src/components/ErrorMessage.css
@@ -1,0 +1,8 @@
+.errorMessage {
+	height: 5rem;
+}
+
+p {
+	margin: 0;
+	padding: 0;
+}
